chore: split memories part 1
