+++ codex-rs/core/src/lib.rs
@@ -57,7 +57,7 @@ pub use codex_mcp::SandboxState;
 mod mcp_openai_file;
 mod mcp_tool_call;
 mod memories;
-pub use memories::clear_memory_roots_contents;
+pub use codex_memories_write::clear_memory_roots_contents;
 pub(crate) mod mention_syntax;
 pub(crate) mod message_history;
 pub(crate) mod utils;
@@ -200,4 +200,5 @@ pub mod compact;
 pub(crate) mod memory_trace;
 pub use memory_trace::BuiltMemory;
 pub use memory_trace::build_memories_from_trace_files;
+mod memory_usage;
 pub mod otel_init;
