+++ codex-rs/core/src/memories/phase2.rs
@@ -1,21 +1,21 @@
 use crate::agent::AgentStatus;
 use crate::agent::status::is_final as is_final_agent_status;
 use crate::config::Config;
-use crate::memories::extensions::prune_old_extension_resources;
-use crate::memories::memory_root;
 use crate::memories::metrics;
 use crate::memories::phase_two;
-use crate::memories::prompts::build_consolidation_prompt;
-use crate::memories::storage::rebuild_raw_memories_file_from_memories;
-use crate::memories::storage::sync_rollout_summaries_from_memories;
-use crate::memories::workspace::memory_workspace_diff;
-use crate::memories::workspace::prepare_memory_workspace;
-use crate::memories::workspace::reset_memory_workspace_baseline;
-use crate::memories::workspace::write_workspace_diff;
 use crate::session::emit_subagent_session_started;
 use crate::session::session::Session;
 use codex_config::Constrained;
 use codex_features::Feature;
+use codex_memories_write::build_consolidation_prompt;
+use codex_memories_write::memory_root;
+use codex_memories_write::prune_old_extension_resources;
+use codex_memories_write::rebuild_raw_memories_file_from_memories;
+use codex_memories_write::sync_rollout_summaries_from_memories;
+use codex_memories_write::workspace::memory_workspace_diff;
+use codex_memories_write::workspace::prepare_memory_workspace;
+use codex_memories_write::workspace::reset_memory_workspace_baseline;
+use codex_memories_write::workspace::write_workspace_diff;
 use codex_protocol::ThreadId;
 use codex_protocol::protocol::AskForApproval;
 use codex_protocol::protocol::SandboxPolicy;
