+++ codex-rs/memories/read/BUILD.bazel
@@ -1,0 +1,9 @@
+load("//:defs.bzl", "codex_rust_crate")
+
+codex_rust_crate(
+    name = "read",
+    crate_name = "codex_memories_read",
+    compile_data = glob([
+        "templates/**",
+    ]),
+)
