+++ codex-rs/memories/read/src/lib.rs
@@ -1,0 +1,19 @@
+//! Read-path helpers for Codex memories.
+//!
+//! This crate owns memory injection, memory citation parsing, and telemetry
+//! classification for read access to the memory folder. It intentionally does
+//! not depend on the memory write pipeline.
+
+pub mod citations;
+mod prompts;
+pub mod usage;
+
+use codex_utils_absolute_path::AbsolutePathBuf;
+
+pub use prompts::build_memory_tool_developer_instructions;
+
+const MEMORY_TOOL_DEVELOPER_INSTRUCTIONS_SUMMARY_TOKEN_LIMIT: usize = 5_000;
+
+pub fn memory_root(codex_home: &AbsolutePathBuf) -> AbsolutePathBuf {
+    codex_home.join("memories")
+}
