+++ codex-rs/memories/read/templates/memories/read_path.md
