+++ codex-rs/memories/write/BUILD.bazel
@@ -1,0 +1,9 @@
+load("//:defs.bzl", "codex_rust_crate")
+
+codex_rust_crate(
+    name = "write",
+    crate_name = "codex_memories_write",
+    compile_data = glob([
+        "templates/**",
+    ]),
+)
