+++ codex-rs/memories/write/src/prompts_tests.rs
@@ -1,9 +1,6 @@
 use super::*;
 use codex_models_manager::model_info::model_info_from_slug;
-use core_test_support::PathExt;
-use pretty_assertions::assert_eq;
 use tempfile::tempdir;
-use tokio::fs as tokio_fs;
 
 #[test]
 fn build_stage_one_input_message_truncates_rollout_using_model_context_window() {
@@ -12,7 +9,7 @@ fn build_stage_one_input_message_truncates_rollout_using_model_context_window() 
     model_info.context_window = Some(123_000);
     let expected_rollout_token_limit = usize::try_from(
         ((123_000_i64 * model_info.effective_context_window_percent) / 100)
-            * phase_one::CONTEXT_WINDOW_PERCENT
+            * STAGE_ONE_CONTEXT_WINDOW_PERCENT
             / 100,
     )
     .unwrap();
@@ -42,7 +39,7 @@ fn build_stage_one_input_message_uses_default_limit_when_model_context_window_mi
     model_info.max_context_window = None;
     let expected_truncated = truncate_text(
         &input,
-        TruncationPolicy::Tokens(phase_one::DEFAULT_STAGE_ONE_ROLLOUT_TOKEN_LIMIT),
+        TruncationPolicy::Tokens(DEFAULT_STAGE_ONE_ROLLOUT_TOKEN_LIMIT),
     );
     let message = build_stage_one_input_message(
         &model_info,
@@ -71,34 +68,4 @@ fn build_consolidation_prompt_points_to_workspace_diff_and_extension_tree() {
         memory_extensions_root.display()
     )));
     assert!(prompt.contains("workspace diff shows deleted extension resource files"));
-}
-
-#[tokio::test]
-async fn build_memory_tool_developer_instructions_renders_embedded_template() {
-    let temp = tempdir().unwrap();
-    let codex_home = temp.path().abs();
-    let memories_dir = codex_home.join("memories");
-    tokio_fs::create_dir_all(&memories_dir).await.unwrap();
-    tokio_fs::write(
-        memories_dir.join("memory_summary.md"),
-        "Short memory summary for tests.",
-    )
-    .await
-    .unwrap();
-
-    let instructions = build_memory_tool_developer_instructions(&codex_home)
-        .await
-        .unwrap();
-
-    assert!(instructions.contains(&format!(
-        "- {}/memory_summary.md (already provided below; do NOT open again)",
-        memories_dir.display()
-    )));
-    assert!(instructions.contains("Short memory summary for tests."));
-    assert_eq!(
-        instructions
-            .matches("========= MEMORY_SUMMARY BEGINS =========")
-            .count(),
-        1
-    );
 }
