+++ codex-rs/memories/write/templates/memories/stage_one_system.md
